feature: ADD db_manager PosgreManager
