+++ src/libs/db_manager.py
@@ -2,6 +2,8 @@ import os
 import pymongo
 from .exception import DBError
 from .error_code import DBErrorCode
+from sqlalchemy.orm import Session
+from sqlalchemy import create_engine
 
 
 class MongoManager:
@@ -17,3 +19,18 @@ class MongoManager:
             return self.client[self.db]
         except Exception as e:
             raise DBError(**DBErrorCode.DBConnectionError.value, err=e)
+
+
+class PostgreManager:
+    def __init__(self) -> None:
+        user: str = os.environ.get('USERNAME')
+        password: str = os.environ.get('PASSWORD')
+        host: str = os.environ.get('POSTGRESQL_HOST')
+        db: str = os.environ.get('DB_NAME')
+        DATABASE_URL = f'postgresql://{user}:{password}@{host}:5432/{db}'
+        self.engine = create_engine(
+            DATABASE_URL, pool_size=5, pool_recycle=100, max_overflow=10
+        )
+
+    def get_session(self):
+        return Session(self.engine)
